Second commit for the ReviewBook logic. Closing book now takes it off the openBook.json file Reviews are updatable as long as book is not closed yet.
There are probably a lot of bugs in there, but that will be for another day.
Also, it is to be noted that once a book is closed, there is now way to view the notes written for this book. Maybe the "NewBook" activity should display another List with the Reviews for the current day. (and element would link to a pure display activity for the review in question). Either that, or something else with a display of the data, per book.